# Enhancement: Optimize has() performance and add a benchmark for has()