@fhein 2018-01-04
I did the benchmarks again testing #220 against develop. Same picture here. Significant performance gain.

	Dumped result to Has.develop.xml
	suite: 133ec88aab16471f81833776daf71e7cd75403da, date: 2018-01-04, stime: 19:37:18
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+
	| benchmark | subject                 | groups | params | revs   | its | mem_peak   | best       | mean       | mode       | worst      | stdev      | rstdev | diff    |
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+
	| HasBench  | benchHasFactory1        |        | []     | 100000 | 100 | 1,313,456b | 0.470020µs | 0.482727µs | 0.480004µs | 0.500030µs | 0.006459µs | 1.34%  | 0.00%   |
	| HasBench  | benchHasInvokable1      |        | []     | 100000 | 100 | 1,313,456b | 0.500020µs | 0.515829µs | 0.510042µs | 0.540030µs | 0.007237µs | 1.40%  | +6.86%  |
	| HasBench  | benchHasService1        |        | []     | 100000 | 100 | 1,313,456b | 0.470030µs | 0.483128µs | 0.480010µs | 0.500030µs | 0.006115µs | 1.27%  | +0.08%  |
	| HasBench  | benchHasAlias1          |        | []     | 100000 | 100 | 1,313,448b | 0.500020µs | 0.518229µs | 0.510199µs | 0.540030µs | 0.008647µs | 1.67%  | +7.35%  |
	| HasBench  | benchHasRecursiveAlias1 |        | []     | 100000 | 100 | 1,313,456b | 0.520020µs | 0.533630µs | 0.530123µs | 0.560040µs | 0.009223µs | 1.73%  | +10.54% |
	| HasBench  | benchHasRecursiveAlias2 |        | []     | 100000 | 100 | 1,313,456b | 0.520030µs | 0.541031µs | 0.540074µs | 0.560040µs | 0.009000µs | 1.66%  | +12.08% |
	| HasBench  | benchHasAbstractFactory |        | []     | 100000 | 100 | 1,313,456b | 0.790040µs | 0.802546µs | 0.800022µs | 0.840050µs | 0.010137µs | 1.26%  | +66.25% |
	| HasBench  | benchHasNot             |        | []     | 100000 | 100 | 1,313,448b | 0.810040µs | 0.829047µs | 0.829534µs | 0.870050µs | 0.010630µs | 1.28%  | +71.74% |
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+

	Dumped result to Has.220.xml
	suite: 133ec88593980ba60d1a68d1cf51a052a3abb242, date: 2018-01-04, stime: 19:26:56
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+
	| benchmark | subject                 | groups | params | revs   | its | mem_peak   | best       | mean       | mode       | worst      | stdev      | rstdev | diff    |
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+
	| HasBench  | benchHasFactory1        |        | []     | 100000 | 100 | 1,315,176b | 0.420020µs | 0.430724µs | 0.430004µs | 0.450030µs | 0.006820µs | 1.58%  | +0.47%  |
	| HasBench  | benchHasInvokable1      |        | []     | 100000 | 100 | 1,315,176b | 0.430020µs | 0.442225µs | 0.440004µs | 0.460030µs | 0.006097µs | 1.38%  | +3.15%  |
	| HasBench  | benchHasService1        |        | []     | 100000 | 100 | 1,315,176b | 0.420020µs | 0.429924µs | 0.430004µs | 0.450030µs | 0.007000µs | 1.63%  | +0.28%  |
	| HasBench  | benchHasAlias1          |        | []     | 100000 | 100 | 1,315,168b | 0.420020µs | 0.428724µs | 0.430004µs | 0.450030µs | 0.005943µs | 1.39%  | 0.00%   |
	| HasBench  | benchHasRecursiveAlias1 |        | []     | 100000 | 100 | 1,315,176b | 0.440020µs | 0.447926µs | 0.450004µs | 0.470030µs | 0.006827µs | 1.52%  | +4.48%  |
	| HasBench  | benchHasRecursiveAlias2 |        | []     | 100000 | 100 | 1,315,176b | 0.440020µs | 0.448625µs | 0.450004µs | 0.470030µs | 0.006786µs | 1.51%  | +4.64%  |
	| HasBench  | benchHasAbstractFactory |        | []     | 100000 | 100 | 1,315,176b | 0.720040µs | 0.732342µs | 0.730060µs | 0.760040µs | 0.007982µs | 1.09%  | +70.82% |
	| HasBench  | benchHasNot             |        | []     | 100000 | 100 | 1,315,168b | 0.770040µs | 0.782644µs | 0.780120µs | 0.820050µs | 0.008789µs | 1.12%  | +82.55% |
	+-----------+-------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+---------+