@fhein 2018-01-04
phpbench does exit with some crude error when I try to use it to compare the dumps. Maybe this is a Windows issue. I'd appreciate if one of the team could check that on Linux.

Find attached all the test parameters and output as well as the according dump-files.

[benchmarks.zip](https://github.com/zendframework/zend-servicemanager/files/1604531/benchmarks.zip)