@fhein 2018-01-09
Closed here because merged with #221.